# apps/frontend/openapi.yaml
openapi: 3.1.0
info:
  title: ALL-Flow API
  version: 1.0.0
  description: |
    ALL-Flow 사내 협업 시스템 백엔드 API 스펙.
    프론트엔드 `lib/types.ts` 와 양방향으로 동기화됩니다.

    인증: Bearer JWT (next-auth v5 발급) — 모든 보호된 엔드포인트에 필요.
servers:
  - url: https://api.allflow.omelet.ai/v1
    description: Production
  - url: http://localhost:8080/api/v1
    description: Local dev
security:
  - bearerAuth: []
...
paths:
...
  /approvals:
    get:
      tags: [approvals]
      summary: 결재 목록
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Approval' } }
    post:
      tags: [approvals]
      summary: 결재 요청 생성
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ApprovalCreate' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Approval' }
  /approvals/{id}/decision:
    post:
      tags: [approvals]
      summary: 결재 의사 결정
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [decision]
              properties:
                decision: { type: string, enum: [approved, rejected] }
                comment: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Approval' }
...
  /channels/{channelId}/messages:
    post:
      tags: [chat]
      summary: 채팅 메시지 전송
      parameters:
        - { name: channelId, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text: { type: string }
      responses:
        '201': { description: Created }
...
  /auth/tokens/revoke:
    post:
      tags: [auth]
      summary: 토큰 폐기 (멱등)
      description: |
        지정된 tokenId 의 폐기를 요청한다. 본 단계는 audit 로그만 기록하며,
        실제 블록리스트 영속화는 후속 작업 (Prisma RevokedToken 모델 + auth 플러그인 검사).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [tokenId]
              properties:
                tokenId: { type: string, minLength: 1, maxLength: 200 }
                reason: { type: string, minLength: 1, maxLength: 500 }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [revoked, tokenId]
                properties:
                  revoked: { type: boolean }
                  tokenId: { type: string }

  # /users/me PATCH merged into /users/me path object above (PDCA-01 cleanup).
...
  /tasks/{id}/comments:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string } }
    get:
      tags: [comments]
      summary: 태스크 코멘트 목록
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Comment' }
    post:
      tags: [comments]
      summary: 태스크 코멘트 추가
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CommentCreate' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Comment' }
  /issues/{id}/comments:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string } }
    get:
      tags: [comments]
      summary: 이슈 코멘트 목록
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Comment' }
    post:
      tags: [comments]
      summary: 이슈 코멘트 추가
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CommentCreate' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Comment' }

  # NOTE: /realtime/ws (WebSocket fallback) is intentionally not declared here.
  # OpenAPI 3.0/3.1 does not standardize WebSocket. Token via query string,
  # protocol: same Notification/Activity events as /realtime/sse.
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Approval:
      type: object
      required: [id, title, requester, approver, status, createdAt]
      properties:
        id: { type: string }
        title: { type: string }
        requester: { type: string }
        approver: { type: string }
        status: { type: string, enum: [pending, approved, rejected, cancelled] }
        amount: { type: number }
        reason: { type: string }
        decidedAt: { type: string, format: date-time }
        createdAt: { type: string, format: date-time }
    ApprovalCreate:
      type: object
      required: [title, approver]
      properties:
        title: { type: string }
        approver: { type: string }
        amount: { type: number }
        reason: { type: string }
...
    Comment:
      type: object
      required: [id, body, author, createdAt]
      properties:
        id: { type: string }
        body: { type: string }
        author:
          type: object
          required: [id, name]
          properties:
            id: { type: string }
            name: { type: string }
        createdAt: { type: string, format: date-time }
    CommentCreate:
      type: object
      required: [body]
      properties:
        body: { type: string, minLength: 1, maxLength: 4000 }